requirements.txt
streamlit==1.32.0
pandas==2.1.4
plotly==5.18.0
psycopg2-binary
google-genai
beautifulsoup4